Fixed narrowing error...

+++ uri.hpp
@@ -33,8 +33,8 @@ namespace uri {
 
     /** Non-owning pointer-size type */
     using Span_t = struct {
-      uint32_t begin;
-      uint32_t end;
+      size_t begin;
+      size_t end;
     };
 
     ///
